chore: Update Firebase Hosting workflow to include working directory
The commit modifies the Firebase Hosting workflows in the `.github/workflows/firebase-hosting-merge.yml` and `.github/workflows/firebase-hosting-pull-request.yml` files. It adds the `working-directory` field with the value `./bs3` to specify the directory where the build command should be run.

This change addresses the need to ensure that the build command is executed in the correct directory, allowing for successful deployment of the project to Firebase Hosting.