Feature/firebase (#3)
* firebase config

* switch syntax

* restored public dir and bug with fab button on mob

* bug with new post form

* VERY permissive firestore rules to get started

* updated boilerplate files

* feat: Add Firebase configuration and Firestore integration

This commit adds the necessary Firebase configuration and Firestore integration to the project. The `firebaseConfig.js` file is created to initialize the Firebase app with the provided configuration. The `db` object is exported to allow interaction with the Firestore database.

The changes in the `post-card.js` file update the import statements to use the new `upVotePost`, `downVotePost`, and `sortPosts` functions from the `firestoreSlice` module instead of the deprecated functions from the `appSlice` module.

The `appSlice.js` file is deleted as it is no longer needed with the introduction of the `firestoreSlice` module.

This commit addresses the need to integrate Firebase and Firestore into the project and update the code to use the new functions from the `firestoreSlice` module.

App loads the posts from firebase and voting works and syncs.

TODO:
- add a new post

* add new post syncs with firestore

* purge credentials and inject via build

* remove env file

* updated CICD

* chore: Update Firebase Hosting workflow names for clarity

The commit updates the names of the Firebase Hosting workflows in the `.github/workflows/firebase-hosting-merge.yml` and `.github/workflows/firebase-hosting-pull-request.yml` files. The `name` field is changed to provide more clarity on the purpose of each workflow.

This change addresses the need to improve the readability and understanding of the workflow names in the repository.

* chore: Update Firebase Hosting workflow to include working directory

The commit modifies the Firebase Hosting workflows in the `.github/workflows/firebase-hosting-merge.yml` and `.github/workflows/firebase-hosting-pull-request.yml` files. It adds the `working-directory` field with the value `./bs3` to specify the directory where the build command should be run.

This change addresses the need to ensure that the build command is executed in the correct directory, allowing for successful deployment of the project to Firebase Hosting.

* chore: Update Firebase Hosting workflows to include environment variables

* revert to native yarn build without env-cmd

* chore: Update Firebase Hosting workflows to include Node.js setup, dependency installation, and build steps

* chore: Update Firebase Hosting workflows to include Node.js setup, dependency installation, and build steps

* fixed CI/CD issues

* chore: Update Firebase Hosting workflows to include Node.js setup, dependency installation, and build steps

* chore: Update Firebase Hosting workflows to use push trigger instead of merge trigger

* Specify working directory

* fixed indentation

* fixed indentation