Feature/edit delete (#23)
* feat: Add error handling to deletePost thunk

The deletePost thunk in firestoreSlice.js has been updated to include error handling. Now, if an error occurs during the deletion of a post, the thunk will reject with the error message. This ensures that any errors are properly handled and can be displayed to the user if necessary.

* chore: Add edit and delete functionality to post card component

* feat: Refactor new post form to include user ID and timestamp

The new post form has been refactored to include the user ID and timestamp in the submitted post. This ensures that each post is associated with the correct user and includes the time of submission.